cluster: split into domain/ and quality/ subpackages
Reorganize the four cluster subprojects under two new parent packages:

  cluster/domain/                domain-style classifiers
          v0/                    (was cluster/v0; the embedding +
                                  K-means clustering pipeline)
          weborganizer/          (was cluster/weborganizer; Dolma3
                                  24-topic classifier)

  cluster/quality/               quality classifiers
          dolma3_quality/        (was cluster/dolma3_quality)
          v0/                    (was cluster/llm_quality; the Sonnet-
                                  distilled fasttext quality model is
                                  the v0 quality classifier)

Mechanical rewrite of every Python import path and every file-system
path referencing the old locations. 32 cluster .py files plus the
store/, scripts/, and docs consumers move in lockstep. No code logic
changes; lint passes.
