+++ experiments/datakit/cluster/domain/v0/ops/exp_smoke.py
@@ -21,7 +21,7 @@ Submit:
     uv run iris --cluster=marin job run --no-wait --cpu=1 --memory=2G \\
         --extra=cpu \\
         --job-name "embed-clusters-smoke-$(date +%Y%m%d-%H%M%S)" \\
-        -- python -m experiments.datakit.cluster.v0.ops.exp_smoke
+        -- python -m experiments.datakit.cluster.domain.v0.ops.exp_smoke
 """
 
 import logging
@@ -39,10 +39,10 @@ from marin.execution.step_runner import StepRunner  # noqa: E402
 from marin.execution.step_spec import StepSpec  # noqa: E402
 from rigging.filesystem import marin_temp_bucket  # noqa: E402
 
-from experiments.datakit.cluster.v0.assign import AssignmentAttrData, assign_source  # noqa: E402
-from experiments.datakit.cluster.v0.sample import sample_centroid_inputs  # noqa: E402
-from experiments.datakit.cluster.v0.summarize import summarize_at_k  # noqa: E402
-from experiments.datakit.cluster.v0.train import train_centroids  # noqa: E402
+from experiments.datakit.cluster.domain.v0.assign import AssignmentAttrData, assign_source  # noqa: E402
+from experiments.datakit.cluster.domain.v0.sample import sample_centroid_inputs  # noqa: E402
+from experiments.datakit.cluster.domain.v0.summarize import summarize_at_k  # noqa: E402
+from experiments.datakit.cluster.domain.v0.train import train_centroids  # noqa: E402
 from experiments.datakit.embeddings.luxical.pipeline import (  # noqa: E402
     LUXICAL_REPO,
     LUXICAL_WEIGHTS_FILE,
