+++ scripts/verify_smoke_v0.py
@@ -50,8 +50,8 @@ from rigging.log_setup import configure_logging
 from zephyr import Dataset, ZephyrContext
 from zephyr.dataset import ShardInfo
 
-from experiments.datakit.cluster.llm_quality.all_sources_quality_llm import LlmQualityOutput
-from experiments.datakit.cluster.v0.assign import AssignmentAttrData
+from experiments.datakit.cluster.domain.v0.assign import AssignmentAttrData
+from experiments.datakit.cluster.quality.v0.all_sources_quality_llm import LlmQualityOutput
 from experiments.datakit.store.all_sources_store import (
     CLUSTER_ASSIGN_ROOT,
     DECONTAM_ROOT,
